Create another JS file
