+++ src/index.js
@@ -1,62 +1,21 @@
 import './style.css';
+import {createListItem} from './createlist.js'
 
-const tasks = [
-  {
-    description: 'Go grocery shopping',
-    completed: false,
-    index: 1,
-  },
-  {
-    description: 'Finish homework',
+const tasks = [];
+
+// Add new task function
+function addTask(description) {
+  const newTask = {
+    description,
     completed: true,
-    index: 2,
-  },
-  {
-    description: 'Take out the trash',
-    completed: false,
-    index: 3,
-  },
-];
-
-function createListItem(task) {
-  const listItem = document.createElement('li');
-  listItem.innerHTML = `
-    <div class="task-item">
-      <div class="task-item__checkbox">
-        <input type="checkbox" id="task-${task.index}" ${task.completed ? 'checked' : ''}>
-        <label for="task-${task.index}"></label>
-      </div>
-      <div class="task-item__description ${task.completed ? 'completed' : ''}">
-        ${task.description}
-      </div>
-      <div class="task-item__actions">
-        <button class="edit-button" data-index="${task.index}">
-          <i class="fas fa-edit"></i>
-        </button>
-        <button class="delete-button" data-index="${task.index}">
-          <i class="fas fa-trash-alt"></i>
-        </button>
-      </div>
-    </div>
-  `;
-
-  const editButton = listItem.querySelector('.edit-button');
-  editButton.addEventListener('click', () => {
-    const descriptionElem = listItem.querySelector('.task-item__description');
-    const inputElem = document.createElement('input');
-    inputElem.type = 'text';
-    inputElem.value = descriptionElem.innerText;
-    inputElem.addEventListener('blur', () => {
-      task.description = inputElem.value;
-      descriptionElem.innerText = inputElem.value;
-    });
-    descriptionElem.innerText = '';
-    descriptionElem.appendChild(inputElem);
-    inputElem.focus();
-  });
-
-  return listItem;
+    index: tasks.length + 1,
+  };
+  tasks.push(newTask);
 }
+addTask('groceries');
+const task = { index: 1, description: 'Task 1', completed: true };
+const listItem = createListItem(task);
+
 function renderTaskList() {
   const taskList = document.getElementById('task-list');
   taskList.innerHTML = '';
@@ -80,20 +39,19 @@ function renderTaskList() {
   `;
   taskList.appendChild(addTaskItem);
 
-  const addTaskForm = document.getElementById('add-task-form');
+const addTaskForm = document.getElementById('add-task-form');
 addTaskForm.addEventListener('submit', (event) => {
   event.preventDefault();
   const taskDescription = document.getElementById('add-task-input').value;
   const newTask = {
     description: taskDescription,
     completed: false, // set the completed property to false
-    index: tasks.length + 1, // set the index property to the value of the new array length
+    index: tasks.length, // set the index property to the value of the new array length
   };
   tasks.push(newTask);
   renderTaskList();
   addTaskForm.reset();
 });
-
 
   tasks.forEach((task) => {
     const listItem = createListItem(task);
@@ -133,5 +91,6 @@ clearButton.addEventListener('click', () => {
       tasks.splice(taskIndex, 1);
     }
   });
+  
   renderTaskList();
 });
